Feat: Add hours and mins, fix month discrepancy
I started on this update by cleaning up the code I had already written.
A lot of the conventions were from TechnoTim, whose tutorial I used to
learn how to make a Twitch bot. I deleted the unnecessary app.js entry
and consolidated code into index.js. After that, I refactored the token
variables into a .env file.

Most of what was changed for this update was in the time-logic.js file.
Previously, I couldn't figure out how to account for discrepancies in
each month's number of days (i.e. 28/30/31). My leap year script was
working pretty well, but I wanted to be as specific as possible when
it came to how many days had passed and how many months that
consequently was.

The solution, I found, was to create an array that contained the number
of days that have passed since Maggie died for each month. So, in May,
30 days have passed since she died. In June, 61 days have passed. In
July, 91 days, so on and so forth. Using the number of days left over
from calculating the number of years that have passed, I was able to
iterate through the array and choose the appropriate number of months
based on the arrays index. Then, I could use those numbers to calculate
days, hours, and minutes.

Finally, I threw all of that into an exported variable and put it into
index.js. I used to have the function saved in a const that was
exported, but this didn't update frequently enough. I'll be testing
regularly to see if the numbers are still accurate in the am.